% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_svg_legend.R
\name{add_svg_legend}
\alias{add_svg_legend}
\title{add a custom svg legend to a leaflet map}
\usage{
add_svg_legend(parent_node, width = 300, height = 100, cols, names, title)
}
\arguments{
\item{parent_node}{an XMLInternalNode}

\item{width}{width of legend in px}

\item{height}{height of legend in px}

\item{cols}{colors of legend}

\item{names}{names/values of colors (text)}

\item{title}{tile to display}
}
\description{
this function is in development
}
\keyword{internal}

