% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/automated_ncdf_update.R
\name{automated_ncdf_update}
\alias{automated_ncdf_update}
\title{Output ncdf file for web service}
\usage{
automated_ncdf_update(output = ".")
}
\arguments{
\item{outpath}{Path to store generated output}
}
\description{
Output ncdf file for web service
}

