% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_web_update.R
\name{automated_web_update}
\alias{automated_web_update}
\title{Wrapper function for automated runs}
\usage{
automated_web_update(outpath = ".")
}
\arguments{
\item{outpath}{Path to store generated output}
}

