% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crn_sites_metadata.R
\name{crn_sites_metadata}
\alias{crn_sites_metadata}
\title{CRN site metadata}
\usage{
crn_sites_metadata(station_ids)
}
\arguments{
\item{station_ids}{Character vector of station ids.}
}
\value{
A data.frame of metadata. All metadata methods return a data.frame with 
\code{station}, \code{latitude}, \code{longitude}, and \code{name}
}
\description{
CRN site metadata for CRN sites.
}
\examples{

crn = crn_data()
meta = crn_sites_metadata(crn$station)

}

