% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moisture_map.R
\name{moisture_map}
\alias{moisture_map}
\title{create a leaflet map for soilmoisture}
\usage{
moisture_map(data, metadata, out_file = NULL)
}
\arguments{
\item{data}{data.frame from data service function (e.g., \link{tx_data})}

\item{metadata}{data.frame from data service metadata function (e.g., \link{ok_sites_metadata})}

\item{out_file}{html output file name for leaflet map}
}
\examples{
\dontrun{
ok = to_hist_percentile(ok_data())
ok_meta = ok_sites_metadata(ok$station)
moisture_map(ok, ok_meta, out_file= '~/soilmoisture_usgs.html')
}
}

