% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ok_sites_metadata.R
\name{ok_sites_metadata}
\alias{ok_sites_metadata}
\title{OK Mesonet site metadata}
\usage{
ok_sites_metadata(station_ids)
}
\arguments{
\item{station_ids}{Character vector of station ids.}
}
\description{
OK site metadata for OK mesonet sites. Data pulled from here
\url{http://www.mesonet.org/index.php/site/sites/station_names_map}
}
\examples{
ok = ok_data()
meta = ok_sites_metadata(ok$station)
}

