% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scan_data.R
\name{scan_data}
\alias{scan_data}
\title{Get Data for a SCAN site}
\usage{
scan_data(stationTriplets, depth, start = Sys.Date() - as.difftime(1, units =
  "days"), end = Sys.Date())
}
\arguments{
\item{stationTriplets}{A character vector of the station ID triplets (e.g., c("9897:NM:UCOOP", "9652:AZ:UCOOP")).
Stations can be queried with \code{\link{scan_sites_by_element_code}}.}

\item{depth}{Individual numeric value for soil moisture sensor depth. In units inches. Common depths include \code{-2, -4, -8, -20, and -40}}

\item{start}{Value with class \code{\link{Date}} for the beginning date for the data query}

\item{end}{Value with class \code{\link{Date}} for the latest date for the data query}
}
\description{

}
\examples{
\dontrun{
#To get pretty much all the data, just do this
stations = scan_sites_by_element_code('SMS')

#most stations seem to have 5 sensors
#2inch sensors
data_2 = scan_data(stations, -2)

#4inch sensors
data_4 = scan_data(stations, -4)

#8 inch sensors
data_8 = scan_data(stations, -8)

#20 inch
data_20 = scan_data(stations, -20)

#40 inch
data_20 = scan_data(stations, -40)
}
}

