% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scan_elements.R
\name{scan_elements}
\alias{scan_elements}
\title{Get all scan database element codes}
\usage{
scan_elements()
}
\value{
a \code{data.frame} with columns elementCd, name, and storedUnitCd
}
\description{
Get all scan database element codes
}
\examples{
all_elements = scan_elements()
print(all_elements$elementCds)
}

