% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_sites_by_element_code.R
\name{scan_sites_by_element_code}
\alias{scan_sites_by_element_code}
\title{Grab all scan sites with certain parameter}
\usage{
scan_sites_by_element_code(elementCd)
}
\arguments{
\item{elementCd}{SCAN element code (see \code{\link{scan_elements}} for full list of codes)}
}
\value{
A list of site triplets as character strings
}
\examples{
\dontrun{
sites_vec = scan_sites_by_element_code('SMS')
}
}

