% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_hist_percentile.R
\name{to_hist_percentile}
\alias{to_hist_percentile}
\title{Convert values to historical percentile}
\usage{
to_hist_percentile(data)
}
\arguments{
\item{data}{A data frame from one of the service functions (e.g., \link{ok_data})}
}

