% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/moisture_map.R
\name{moisture_map}
\alias{moisture_map}
\title{create a leaflet map for soilmoisture}
\usage{
moisture_map(data, metadata, out_file)
}
\arguments{
\item{data}{data.frame from data service function (e.g., \link{tx_data})}

\item{metadata}{data.frame from data service metadata function (e.g., \link{ok_sites_metadata})}

\item{out_file}{html output file name for leaflet map}
}
\description{
create a leaflet map for soilmoisture
}
\examples{
\dontrun{
moisture_map(to_hist_percentile(ok_data()), ok_sites_metadata(ok_data()$station), out_file= 'soilmoisture_usgs.html')
}
}

