% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/moisture_map.R
\name{moisture_map}
\alias{moisture_map}
\title{create a leaflet map for soilmoisture}
\usage{
moisture_map(depth, out_file = "soilmoisture_usgs.html")
}
\arguments{
\item{depth}{in inches for sensors}

\item{out_file}{html output file name for leaflet map}
}
\description{
create a leaflet map for soilmoisture
}
\examples{
\dontrun{
moisture_map(-2)
}
}

