% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scan_sites_metadata.R
\name{scan_sites_metadata}
\alias{scan_sites_metadata}
\title{Get SCAN station metadata}
\usage{
scan_sites_metadata(stationTriplets)
}
\arguments{
\item{stationTriplets}{A character vector of the station ID triplets (e.g., c("9897:NM:UCOOP", "9652:AZ:UCOOP")).
Stations can be queried with \code{\link{scan_sites_by_element_code}}.}
}
\value{
data frame with a row for each station. Order returned same as requested station order is not guaranteed.
}
\description{
Get SCAN station metadata
}
\note{
This is somewhat slow. Queries over 1000 sites are split into multiple requests.
I recommend querying only the specific sites you need.
}
\examples{
scan_sites_metadata("9897:NM:UCOOP")
}
\author{
Luke Winslow
}

